add tests, cleanup
